detal+verif done
