+++ front-end/src/components/navbar.jsx
@@ -1,118 +1,204 @@
+import { Register } from "./register";
+
 import {
-    Box,
-    Flex,
-    Text,
-    IconButton,
-    Button,
-    Stack,
-    Collapse,
-    Icon,
-    Popover,
-    PopoverTrigger,
-    PopoverContent,
-    useColorModeValue,
-    useBreakpointValue,
-    useDisclosure,
-  } from "@chakra-ui/react";
-  import {Link,useNavigate} from "react-router-dom"
-  import {
-    HamburgerIcon,
-    CloseIcon,
-    ChevronDownIcon,
-    ChevronRightIcon,
-  } from "@chakra-ui/icons";
-  import {Register} from "../components/register"
+  Box,
+  Flex,
+  Text,
+  IconButton,
+  Button,
+  Stack,
+  Collapse,
+  Icon,
+  Link,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  useColorModeValue,
+  useDisclosure,
+  useBreakpointValue,
+  useColorMode,
+  Image,
+  FormControl,
+  Input,
+  Menu,
+  MenuDivider,
+  MenuButton,
+  Avatar,
+  MenuList,
+  MenuItem,
+} from "@chakra-ui/react";
+import {
+  HamburgerIcon,
+  CloseIcon,
+  ChevronDownIcon,
+  ChevronRightIcon,
+  MoonIcon,
+  SunIcon,
+} from "@chakra-ui/icons";
+import { Link as Link2 } from "react-router-dom";
+import NavbarLogin from "./login";
 
-  import { ImLibrary } from "react-icons/im";
+import { useDispatch, useSelector } from "react-redux";
+import { logout, login } from "../redux/userSlice";
+import { useRef, useState, useEffect } from "react";
+import Swal from "sweetalert2";
+import Axios from "axios";
+import { useNavigate } from "react-router-dom";
 
-  export default function Navbar() {
-    const { isOpen, onToggle } = useDisclosure();
-    const navigate = useNavigate()
-    
-    return (
-      <Box>
+export default function NavbarComp() {
+  const { NIM, username, email, isVerified } = useSelector(
+    (state) => state.userSlice.value
+  );
+  console.log(NIM, username, isVerified);
+  const { isOpen, onToggle, onClose, onOpen } = useDisclosure();
+  //   const { colorMode, toggleColorMode } = useColorMode();
+  const tokenlocalstorage = localStorage.getItem("token");
+  const dispatch = useDispatch();
+
+  let [token, setToken] = useState("");
+  let navigate = useNavigate();
+
+  const onLogout = () => {
+    dispatch(logout());
+    localStorage.removeItem("token");
+  };
+
+  const onVerification = async () => {
+    try {
+      const result = await Axios.post("http://localhost:2000/users/changeotp", {
+        NIM,
+      });
+      setToken(result.data.token);
+      Swal.fire({
+        icon: "success",
+        title: "Your verification code already sent to your email.",
+        text: `${result.data.message}`,
+        timer: 2000,
+        customClass: {
+          container: "my-swal",
+        },
+      });
+      setTimeout(() => navigate(`/verification/${result.data.token}`), 2000);
+    } catch (err) {
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: `${err.response.data}`,
+        customClass: {
+          container: "my-swal",
+        },
+      });
+    }
+  };
+
+  return (
+    <Box>
+      <Flex
+        bg={useColorModeValue("white", "gray.800")}
+        color={useColorModeValue("gray.600", "white")}
+        minH={"60px"}
+        py={{ base: 2 }}
+        px={{ base: 4 }}
+        borderBottom={1}
+        borderStyle={"solid"}
+        borderColor={useColorModeValue("gray.200", "gray.900")}
+        align={"center"}
+      >
         <Flex
-          bg={useColorModeValue("white", "gray.800")}
-          color={useColorModeValue("gray.600", "white")}
-          minH={"60px"}
-          py={{ base: 2 }}
-          px={{ base: 4 }}
-          borderBottom={1}
-          borderStyle={"solid"}
-          borderColor={useColorModeValue("gray.200", "gray.900")}
-          align={"center"}
+          as={Link2}
+          to="/"
+          flex={{ base: 2 }}
+          justify={{ md: "start" }}
+          align="center"
         >
-          <Flex
-            flex={{ base: 1, md: "auto" }}
-            ml={{ base: -2 }}
-            display={{ base: "flex", md: "none" }}
+          <Image
+            src="https://cdn-icons-png.flaticon.com/512/2997/2997509.png"
+            w="8"
+          />
+          <Text
+            textAlign={useBreakpointValue({ base: "center", md: "left" })}
+            fontFamily={"heading"}
+            color={useColorModeValue("gray.800", "white")}
           >
-            
-            <IconButton
-              onClick={onToggle}
-              icon={
-                isOpen ? <CloseIcon w={3} h={3} /> : <HamburgerIcon w={5} h={5} />
-              }
-              variant={"ghost"}
-              aria-label={"Toggle Navigation"}
-            />
+            LibraryID
+          </Text>
+          <Flex display={{ base: "none", md: "flex" }} ml={10}>
+            <DesktopNav />
+          </Flex>
+        </Flex>
 
-          </Flex>
-          <Flex flex={{ base: 1 }} justify={{ base: "center", md: "start" }}>
-          <Flex as={Link} to = "/" >
-          <Icon
-                    fontSize="xl"
-                    as={ImLibrary}
-                    sx={{ _hover: { cursor: "pointer" } }}                />
-            <Text
-              textAlign={useBreakpointValue({ base: "center", md: "left" })}
-              fontFamily={"heading"}
-              color={useColorModeValue("gray.800", "white")}
+        {/* <Button
+          onClick={toggleColorMode}
+          bg={useColorModeValue("white", "gray.800")}
+        >
+          {colorMode === "light" ? <MoonIcon /> : <SunIcon />}
+        </Button> */}
+
+        {tokenlocalstorage ? (
+          <Menu>
+            <MenuButton
+              display={{ base: "none", md: "inline-flex" }}
+              as={Button}
+              rounded={"full"}
+              cursor={"pointer"}
+              minW={0}
+              p="0"
             >
-              LibraryID
-            </Text>
-          </Flex>
-  
-            <Flex display={{ base: "none", md: "flex" }} ml={10}>
-              <DesktopNav />
-            </Flex>
-          </Flex>
-  
+              <Flex>
+                <Avatar name={username} size="md" />
+              </Flex>
+            </MenuButton>
+            <MenuList alignItems={"center"}>
+              <MenuItem>{username}</MenuItem>
+              <MenuItem>{NIM}</MenuItem>
+              <MenuDivider />
+              <MenuItem>Profile</MenuItem>
+              {isVerified ? (
+                ""
+              ) : (
+                <MenuItem onClick={onVerification}>Verifcation Account</MenuItem>
+              )}
+              <MenuItem as={Link2} to="/" onClick={onLogout}>
+                Log Out
+              </MenuItem>
+            </MenuList>
+          </Menu>
+        ) : (
           <Stack
-            flex={{ base: 1, md: 0 }}
+            display={{ base: "none", md: "inline-flex" }}
             justify={"flex-end"}
             direction={"row"}
-            spacing={6}
           >
-            <Button
-              as={Link}
-              fontSize={"sm"}
-              fontWeight={400}
-              variant={"link"}
-              href={"#"}
-              to= "/login"
-            >
-              Sign In
-            </Button>
-
-              <Register/>
-           
+            <NavbarLogin />
+            <Register />
           </Stack>
+        )}
+        <Flex display={{ base: "flex", md: "none" }}>
+          <IconButton
+            onClick={onToggle}
+            icon={
+              isOpen ? <CloseIcon w={3} h={3} /> : <HamburgerIcon w={5} h={5} />
+            }
+            variant={"ghost"}
+            aria-label={"Toggle Navigation"}
+          />
         </Flex>
-  
-        <Collapse in={isOpen} animateOpacity>
-          <MobileNav />
-        </Collapse>
-      </Box>
-    );
-  }
-  
-  const DesktopNav = () => {
-    const linkColor = useColorModeValue("gray.600", "gray.200");
-    const linkHoverColor = useColorModeValue("gray.800", "white");
-    const popoverContentBgColor = useColorModeValue("white", "gray.800");
-  
-    return (
+      </Flex>
+
+      <Collapse in={isOpen} animateOpacity>
+        <MobileNav />
+      </Collapse>
+    </Box>
+  );
+}
+
+const DesktopNav = () => {
+  const linkColor = useColorModeValue("gray.600", "gray.200");
+  const linkHoverColor = useColorModeValue("gray.800", "white");
+  const popoverContentBgColor = useColorModeValue("white", "gray.800");
+
+  return (
+    <>
       <Stack direction={"row"} spacing={4}>
         {NAV_ITEMS.map((navItem) => (
           <Box key={navItem.label}>
@@ -132,7 +218,7 @@ import {
                   {navItem.label}
                 </Link>
               </PopoverTrigger>
-  
+
               {navItem.children && (
                 <PopoverContent
                   border={0}
@@ -153,64 +239,178 @@ import {
           </Box>
         ))}
       </Stack>
-    );
-  };
-  
-  const DesktopSubNav = ({ label, href, subLabel }) => {
-    return (
-      <Link
-        href={href}
-        role={"group"}
-        display={"block"}
-        p={2}
-        rounded={"md"}
-        _hover={{ bg: useColorModeValue("pink.50", "gray.900") }}
-      >
-        <Stack direction={"row"} align={"center"}>
-          <Box>
-            <Text
-              transition={"all .3s ease"}
-              _groupHover={{ color: "pink.400" }}
-              fontWeight={500}
-            >
-              {label}
-            </Text>
-            <Text fontSize={"sm"}>{subLabel}</Text>
-          </Box>
-          <Flex
+    </>
+  );
+};
+
+const DesktopSubNav = ({ label, href, subLabel }) => {
+  return (
+    <Link
+      href={href}
+      role={"group"}
+      display={"block"}
+      p={2}
+      rounded={"md"}
+      _hover={{ bg: useColorModeValue("pink.50", "gray.900") }}
+    >
+      <Stack direction={"row"} align={"center"}>
+        <Box>
+          <Text
             transition={"all .3s ease"}
-            transform={"translateX(-10px)"}
-            opacity={0}
-            _groupHover={{ opacity: "100%", transform: "translateX(0)" }}
-            justify={"flex-end"}
-            align={"center"}
-            flex={1}
+            _groupHover={{ color: "pink.400" }}
+            fontWeight={500}
           >
-            <Icon color={"pink.400"} w={5} h={5} as={ChevronRightIcon} />
-          </Flex>
-        </Stack>
-      </Link>
-    );
+            {label}
+          </Text>
+          <Text fontSize={"sm"}>{subLabel}</Text>
+        </Box>
+        <Flex
+          transition={"all .3s ease"}
+          transform={"translateX(-10px)"}
+          opacity={0}
+          _groupHover={{ opacity: "100%", transform: "translateX(0)" }}
+          justify={"flex-end"}
+          align={"center"}
+          flex={1}
+        >
+          <Icon color={"blue.400"} w={5} h={5} as={ChevronRightIcon} />
+        </Flex>
+      </Stack>
+    </Link>
+  );
+};
+
+const MobileNav = () => {
+  const tokenlocalstorage = localStorage.getItem("token");
+  const { NIM, username, email, isVerified } = useSelector(
+    (state) => state.userSlice.value
+  );
+  const dispatch = useDispatch();
+  // const inputNIM = useRef("");
+  // const inputPASS = useRef("");
+  let [token, setToken] = useState("");
+
+  const onLogout = () => {
+    dispatch(logout());
+    localStorage.removeItem("token");
   };
-  
-  const MobileNav = () => {
-    return (
+
+  //   const onLogin = async () => {
+  //     try {
+  //       const user = {
+  //         password: inputPASS.current.value,
+  //         NIM: inputNIM.current.value,
+  //       };
+
+  //       console.log(user);
+
+  //       const result = await Axios.post(url, user);
+
+  //       dispatch(
+  //         login({
+  //           NIM: result.data.isUserExist.NIM,
+  //           username: result.data.isUserExist.username,
+  //           email: result.data.isUserExist.email,
+  //         })
+  //       );
+
+  //       localStorage.setItem("token", result.data.token);
+  //     } catch (err) {
+  //       Swal.fire({
+  //         icon: "error",
+  //         title: "Oops...",
+  //         text: `${err.response.data}`,
+  //         timer: 1000,
+  //         customClass: {
+  //           container: "my-swal",
+  //         },
+  //       });
+  //     }
+  //   };
+
+  const onVerification = async () => {
+    try {
+      const result = await Axios.post("http://localhost:2000/users/changeotp", {
+        NIM,
+      });
+      setToken(result.data.token);
+      Swal.fire({
+        icon: "success",
+        title: "Good Job",
+        text: `${result.data.massage}`,
+        timer: 2000,
+        customClass: {
+          container: "my-swal",
+        },
+      });
+      // setTimeout(() => setMove(true), 2000);
+    } catch (err) {
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: `${err.response.data}`,
+        customClass: {
+          container: "my-swal",
+        },
+      });
+    }
+  };
+
+  return (
+    <>
       <Stack
         bg={useColorModeValue("white", "gray.800")}
-        p={4}
+        pr={4}
+        pl={4}
         display={{ md: "none" }}
       >
         {NAV_ITEMS.map((navItem) => (
           <MobileNavItem key={navItem.label} {...navItem} />
         ))}
+
+        {tokenlocalstorage ? (
+          <Menu>
+            <MenuButton>
+              <Flex>
+                <Avatar size="sm" name={username} />
+                {/* <Box ml="3">
+                  <Text fontWeight="bold">
+                    {username} {NIM}
+                  </Text>
+                </Box> */}
+                <MenuItem>{username}</MenuItem>
+                <MenuItem>{NIM}</MenuItem>
+              </Flex>
+            </MenuButton>
+
+            <MenuList alignItems={"center"}>
+              <MenuItem>Profile</MenuItem>
+              <MenuItem onClick={onLogout}>Log Out</MenuItem>
+              {isVerified ? (
+                ""
+              ) : (
+                <MenuItem onClick={onVerification}>Verification Account</MenuItem>
+              )}
+            </MenuList>
+          </Menu>
+        ) : (
+          <Stack justify={"flex-end"}>
+            <Stack direction="row">
+              <NavbarLogin/>
+              <Register />
+            </Stack>
+          </Stack>
+        )}
       </Stack>
-    );
-  };
-  
-  const MobileNavItem = ({ label, children, href }) => {
-    const { isOpen, onToggle } = useDisclosure();
-  
-    return (
+    </>
+  );
+};
+
+const MobileNavItem = ({ label, children, href }) => {
+  const { isOpen, onToggle } = useDisclosure();
+
+  return (
+    <>
       <Stack spacing={4} onClick={children && onToggle}>
         <Flex
           py={2}
@@ -238,10 +438,9 @@ import {
             />
           )}
         </Flex>
-  
-        <Collapse in={isOpen} animateOpacity style={{ marginTop: "0!important" }}>
+
+        <Collapse in={isOpen} animateOpacity>
           <Stack
-            mt={2}
             pl={4}
             borderLeft={1}
             borderStyle={"solid"}
@@ -250,54 +449,46 @@ import {
           >
             {children &&
               children.map((child) => (
-                <Link key={child.label} py={2} href={child.href}>
+                <Link key={child.label} href={child.href}>
                   {child.label}
                 </Link>
               ))}
           </Stack>
         </Collapse>
       </Stack>
-    );
-  };
-  
-  const NAV_ITEMS = [
-    {
-      label: "Inspiration",
-      children: [
-        {
-          label: "Explore Design Work",
-          subLabel: "Trending Design to inspire you",
-          href: "#",
-        },
-        {
-          label: "New & Noteworthy",
-          subLabel: "Up-and-coming Designers",
-          href: "#",
-        },
-      ],
-    },
-    {
-      label: "Find Work",
-      children: [
-        {
-          label: "Job Board",
-          subLabel: "Find your dream design job",
-          href: "#",
-        },
-        {
-          label: "Freelance Projects",
-          subLabel: "An exclusive list for contract work",
-          href: "#",
-        },
-      ],
-    },
-    {
-      label: "Learn Design",
-      href: "#",
-    },
-    {
-      label: "Hire Designers",
-      href: "#",
-    },
-  ];
-  
+    </>
+  );
+};
+
+const NAV_ITEMS = [
+  {
+    label: "Inspiration",
+    children: [
+      {
+        label: "Explore Design Work",
+        subLabel: "Trending Design to inspire you",
+        href: "/",
+      },
+      {
+        label: "New & Noteworthy",
+        subLabel: "Up-and-coming Designers",
+        href: "/",
+      },
+    ],
+  },
+  {
+    label: "Find Work",
+    children: [
+      {
+        label: "Job Board",
+        subLabel: "Find your dream design job",
+        href: "/",
+      },
+      {
+        label: "Freelance Projects",
+        subLabel: "An exclusive list for contract work",
+        href: "/",
+      },
+    ],
+  },
+];
