+++ front-end/src/components/register.jsx
@@ -60,10 +60,10 @@ export const Register = () => {
 
       dispatch(
         login({
-          NIM: result.data.NIM,
-          username: result.data.username,
-          email: result.data.email,
-          isVerified: result.data.isVerified,
+          NIM: result.data.data.NIM,
+          username: result.data.data.username,
+          email: result.data.data.email,
+            isVerified: result.data.data.isVerified,
 
         })
       );
@@ -206,7 +206,7 @@ export const Register = () => {
                           </FormControl>
                           <ModalFooter>
                             <Button type="submit" colorScheme="blue" mr={3}>
-                              Sign up
+                              Register
                             </Button>
                             <Button onClick={onClose}>Cancel</Button>
                           </ModalFooter>
